chore-optimization-jobeomjun--함수명 설정
