+++ src/components/shose-search/itemlistbox/itemcard/itemcard.css.ts
