+++ src/components/shose-search/itemlistbox/itemlistbox.css.ts
