+++ src/components/text-search/mainContainder/maincontainer.css.ts
