+++ src/router/index.tsx
@@ -1,7 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
 import App from '../App';
-import ImageSearch from '../components/choose-shose/image-search/ImageSearch';
-import TextSearch from '../components/choose-shose/text-search/TextSearch';
 import EmptyShoesRoom from '../components/empty-shoes-room/EmptyShoesRoom';
 import FootInfo from '../components/foot-info/FootInfo';
 import Login from '../components/login/Login';
@@ -19,6 +17,8 @@ import Mypage from '../pages/mypage/Mypage';
 import NotFound from '../pages/not-found/notFound';
 import SharePage from '../pages/share-page/SharePage';
 import ViewedHistoryPage from '../pages/viewed-history-page/ViewedHistoryPage';
+import ImageSearch from '../pages/imagesearch-page/ImageSearch';
+import TextSearch from '../pages/textsearch-page/TextSearch';
 
 export const router = createBrowserRouter([
   {
